✨ *refact(decorators): reorganizar e adicionar novos decoradores para tarefas*
Este commit reorganiza a estrutura dos decoradores e adiciona novos decoradores para melhorar a funcionalidade das tarefas compartilhadas.

Files Changed:
- *crawjud_app/_wrapper.py*: remoção do arquivo obsoleto que continha decoradores.
- *crawjud_app/decorators/__init__.py*: criação de um novo módulo para fornecer decoradores tipados para tarefas compartilhadas do Celery.
- *crawjud_app/decorators/_bot.py*: adição de novos decoradores para inicialização e execução de classes de bot.
- *crawjud_app/bots/pje/capa.py*: atualização do import para o novo caminho do módulo de decoradores.
