+++ crawjud_app/bots/pje/capa.py
@@ -11,11 +11,11 @@ from typing import TYPE_CHECKING
 from dotenv import load_dotenv
 from httpx import Client
 
-from crawjud.wrapper import wrap_cls
 from crawjud_app._wrapper import shared_task
 from crawjud_app.bots.pje._controler import PjeBot
 from crawjud_app.bots.resources.formatadores import formata_tempo
 from crawjud_app.custom._task import ContextTask
+from crawjud_app.decorators.bot import wrap_cls
 from crawjud_app.types.pje import DictResults
 
 if TYPE_CHECKING:
